toolkit/components/passwordmgr/test/mochitest/test_autocomplete_hasBeenTypePassword.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test that passwords are autocompleted into fields that were previously type=password</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: Test that passwords are autocompleted into fields that were previously type=password

<script>
var setupScript = runInParent(function setup() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  // Add two logins to prevent autofill on page load.
  let nsLoginInfo = Components.Constructor("@mozilla.org/login-manager/loginInfo;1",
                                           Ci.nsILoginInfo, "init");
  assert.ok(nsLoginInfo != null, "nsLoginInfo constructor");

  let login1 = new nsLoginInfo("https://example.com", "https://example.com", null,
                               "user1", "pass1");
  let login2 = new nsLoginInfo("https://example.com", "https://example.com", null,
                               "user2", "pass2");

  Services.logins.addLogin(login1);
  Services.logins.addLogin(login2);
});
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 tests multiple matching logins -->
  <form id="form1" action="https://example.com/formtest.js" onsubmit="return false;">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">
    <button type="submit">Submit</button>
  </form>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var uname = $_(1, "uname");
var pword = $_(1, "pword");

// Restore the form to the default state.
function restoreForm() {
  uname.value = "";
  pword.value = "";
  uname.focus();
}

function spinEventLoop() {
  return Promise.resolve();
}

add_task(async function setup() {
  listenForUnexpectedPopupShown();
});

add_task(async function test_form1_initial_empty() {
  await SimpleTest.promiseFocus(window);

  // Make sure initial form is empty.
  checkLoginForm(uname, "", pword, "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

add_task(async function test_form1_password_to_type_text() {
  await SimpleTest.promiseFocus(window);
  info("Setting the password field type to text");
  // This is similar to a site implementing their own password visibility/unmasking toggle
  pword.type = "text";

  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  checkAutoCompleteResults(results, ["user1", "user2"], "example.com", "Check all menuitems are displayed correctly.");

  synthesizeKey("KEY_ArrowDown"); // first item
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting

  synthesizeKey("KEY_Enter");

  await promiseFormsProcessed();
  is(uname.value, "user1", "username should match the login, not the password");
  is(pword.value, "pass1", "password should match the login, not the username");
  checkLoginForm(uname, "user1", pword, "pass1");

  restoreForm();
  info("Focusing the password field");
  pword.focus();

  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // first item
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting

  synthesizeKey("KEY_Enter");
  await spinEventLoop();
  is(pword.value, "pass1", "Password should match the login that was selected");
  checkLoginForm(uname, "", pword, "pass1");
});
</script>
</pre>
</body>
</html>
